chore: comprehensive code cleanup

+++ src/components/ItemPopUpHeader.tsx
@@ -1,21 +1,50 @@
 import React, { memo } from 'react';
 import { StyleSheet, TextInput, Alert } from 'react-native';
-import { Icon } from 'react-native-elements'
+import { Icon } from 'react-native-elements';
 import { useData } from './DataProvider';
 import * as FileSystem from 'expo-file-system';
 import { colors } from '../utils/constants';
+import { ClothingItem, ExistingClothingItem } from '../utils/schemas';
 
-const ItemPopUpHeader = (props: any) => {
-  const {
-    item,
-    navigation,
-    onChangeText,
-    editable,
-    newItem,
-  } = props;
+interface ItemPopUpHeaderProps {
+  // item to be displayed, edited, deleted
+  item?: ClothingItem;
+
+  // function that navigates back to the previous screen
+  goBack: () => void;
+
+  // function that sets the name of the item
+  onChangeText: (text: string) => void;
+
+  // whether or not the item can be edited
+  editable: boolean;
+}
+
+/**
+ * Header component designed for the ItemDetailsPage modal. Allows for changing
+ * the name of and deletion of an item.
+ *
+ * @component
+ * @param props {ItemPopUpHeaderProps}
+ * @returns {JSX.Element} the ItemPopUpHeader component.
+ * @example
+ * return (
+ *   <ItemPopUpHeader editable={true} item={item} goBack={goBack} onChangeText={onChangeText}/>
+ * )
+ */
+const ItemPopUpHeader = (props: ItemPopUpHeaderProps) => {
+  const { item = {}, goBack, onChangeText, editable } = props;
 
   const { deleteItem } = useData();
 
+  // Type guard to check if item is an ExistingClothingItem
+  function isExistingClothingItem(
+    item: ClothingItem,
+  ): item is ExistingClothingItem {
+    return 'id' in item;
+  }
+
+  // Brings up pop-up alert confirming deletion of item
   const deleteOnPress = () => {
     Alert.alert(
       'Delete Item',
@@ -31,41 +60,45 @@ const ItemPopUpHeader = (props: any) => {
             if (item?.imageUrl) {
               await FileSystem.deleteAsync(item.imageUrl);
             }
-            await deleteItem(item.id);
-            navigation.goBack();
+            if (isExistingClothingItem(item)) {
+              await deleteItem(item.id);
+            }
+            goBack();
           },
           style: 'destructive',
         },
       ],
-      { cancelable: true }
+      { cancelable: true },
     );
   };
 
   return (
     <>
       <Icon
-        name='close'
-        type='material'
+        name="close"
+        type="material"
         color={colors.foreground}
         size={35}
         containerStyle={styles.exitIcon}
-        onPress={() => navigation.goBack()}
+        onPress={() => goBack()}
       />
       <TextInput
         style={styles.textInput}
-        value={item?.name}
+        value={item.name}
         onChangeText={onChangeText}
         placeholder="NEW ITEM"
         editable={editable}
-      /> 
-      {!newItem &&editable && <Icon
-        name='delete'
-        type='material'
-        color={colors.foreground}
-        size={35}
-        containerStyle={styles.deleteIcon}
-        onPress={deleteOnPress}
-      />}
+      />
+      {isExistingClothingItem(item) && editable && (
+        <Icon
+          name="delete"
+          type="material"
+          color={colors.foreground}
+          size={35}
+          containerStyle={styles.deleteIcon}
+          onPress={deleteOnPress}
+        />
+      )}
     </>
   );
 };
@@ -78,11 +111,11 @@ const styles = StyleSheet.create({
   },
   textInput: {
     width: 239,
-    height: 28, 
-    textAlign: 'center', 
-    color: colors.foreground, 
-    fontSize: 24, 
-    fontWeight: '400', 
+    height: 28,
+    textAlign: 'center',
+    color: colors.foreground,
+    fontSize: 24,
+    fontWeight: '400',
     letterSpacing: 2,
     paddingTop: 5,
     marginBottom: 20,
